CalcDistance moved to utils folder and imported in
